now Vector have double components, add game objects

+++ RocketWanderer/WpfApp/Views/Menu/MenuViewWpf.cs
@@ -60,8 +60,8 @@ namespace WpfApp.Views.Menus
       _stackPanel.Width = menuSize.X;
       _stackPanel.Height = menuSize.Y;
 
-      int leftOffset = parParentSize.X / 2 - menuSize.X / 2;
-      int topOffset  = parParentSize.Y / 2 - menuSize.Y / 2 + parParentSize.Y / 20;
+      double leftOffset = parParentSize.X / 2 - menuSize.X / 2;
+      double topOffset  = parParentSize.Y / 2 - menuSize.Y / 2 + parParentSize.Y / 20;
 
       System.Windows.Controls.Canvas.SetLeft(_stackPanel, leftOffset);
       System.Windows.Controls.Canvas.SetTop(_stackPanel, topOffset);
